Move tests to XUnit
